:sparkles: `errors` Improve the error that is returned when a response cannot be unmarshalled as JSON (#104)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Proprietary
-->
### Description

<!--
Please add any detail or context that would be useful to a reviewer.
-->

Improve the error that is returned when a response cannot be
unmarshalled as JSON

### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [x]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).