⚙️ Add a check in CallAndCheckSuccess for response unmarshalling issues (#107)
<!--
Copyright (C) 2020-2022 Arm Limited or its affiliates and Contributors.
All rights reserved.
SPDX-License-Identifier: Proprietary
-->
### Description
This changes add an extra check CallAndCheckSuccess to detect empty
unmarshalled responses.
The restriction in the return type [T] to be a pointer allows us to
optimize calling reflection.IsEmpty only when that pointer is not nil


### Test Coverage

<!--
Please put an `x` in the correct box e.g. `[x]` to indicate the testing
coverage of this change.
-->

- [X]  This change is covered by existing or additional automated tests.
- [ ] Manual testing has been performed (and evidence provided) as
automated testing was not feasible.
- [ ] Additional tests are not required for this change (e.g.
documentation update).